+++ src/graph/dijkstra/problem.cpp
@@ -5,19 +5,19 @@
 
 class Solution {
 public:
-    static int lc_1334(int n, std::vector<std::vector<int>>& edges, int distanceThreshold) {
-        std::vector<std::vector<std::pair<int, int>>> dct(n);
+    static long long lc_1334(long long n, std::vector<std::vector<long long>>& edges, long long distanceThreshold) {
+        std::vector<std::vector<std::pair<long long, long long>>> dct(n);
         for (const auto& edge : edges) {
-            int i = edge[0];
-            int j = edge[1];
-            int w = edge[2];
+            long long i = edge[0];
+            long long j = edge[1];
+            long long w = edge[2];
             dct[i].emplace_back(j, w);
             dct[j].emplace_back(i, w);
         }
 
         float res = std::numeric_limits<float>::max();
-        int ans = 0;
-        for (int i = 0; i < n; ++i) {
+        long long ans = 0;
+        for (long long i = 0; i < n; ++i) {
             std::vector<float> dis = Dijkstra::get_shortest_path(dct, i);
             float cur = 0;
             for (float distance : dis) {
