+++ tests/leetcode/problem_1.cpp
@@ -10,30 +10,30 @@
 
 class SnapshotArray {
 public:
-    SnapshotArray(int length) {
+    SnapshotArray(long long length) {
         snap_cnt = 0;
-        lst = std::vector<std::vector<int>> (length, {0});
-        snap_ids = std::vector<std::vector<int>> (length, {0});
+        lst = std::vector<std::vector<long long>> (length, {0});
+        snap_ids = std::vector<std::vector<long long>> (length, {0});
 
     }
 
-    void set(int index, int val) {
+    void set(long long index, long long val) {
         lst[index].push_back(val);
         snap_ids[index].push_back(val);
     }
 
-    int snap() {
+    long long snap() {
         snap_cnt++;
         return snap_cnt - 1;
     }
 
-    int get(int index, int snap_id) {
-        int i = std::upper_bound(snap_ids[index].begin(), snap_ids[index].end(), snap_id) - snap_ids[index].begin();
-        int ans = lst[index][i-1];
+    long long get(long long index, long long snap_id) {
+        long long i = std::upper_bound(snap_ids[index].begin(), snap_ids[index].end(), snap_id) - snap_ids[index].begin();
+        long long ans = lst[index][i-1];
         return ans;
     }
 private:
-    int snap_cnt=0;
-    std::vector<std::vector<int>> lst;
-    std::vector<std::vector<int>> snap_ids;
+    long long snap_cnt=0;
+    std::vector<std::vector<long long>> lst;
+    std::vector<std::vector<long long>> snap_ids;
 };
